Change meter system

+++ app/Http/Controllers/LifelineController.php
@@ -77,6 +77,12 @@ class LifelineController extends Controller
         if ($request->date_of_application < '2023-01-01' || $request->date_of_application == $request->date_of_birth) {
             return redirect()->back()->withError('Pls double Check the Date of Application!');
         }
+
+        // check if account number existing in the record
+        $lifeline = Lifeline::where('account_no', $request->electric_service_details)->first();
+        if($lifeline){
+            return redirect()->back()->withError('Account No. already existing in the record!');
+        }
         
         $year = date("Y");
         $control_id = Helper::IDGenerator(new Lifeline, 'control_no', 4, $year); /** Generate control no */
@@ -138,6 +144,16 @@ class LifelineController extends Controller
             'annual_income' => ['required'],
             'sdwo_certification' => ['required', 'string', 'max:255'],
         ]);
+
+        if ($request->date_of_application < '2023-01-01' || $request->date_of_application == $request->date_of_birth) {
+            return redirect()->back()->withError('Pls double Check the Date of Application!');
+        }
+
+        // check if account number existing in the record
+        $lifeline = Lifeline::where('account_no', $request->electric_service_details)->first();
+        if($lifeline){
+            return redirect()->back()->withError('Account No. already existing in the record!');
+        }
 
         $year = date("Y");
         $control_id = Helper::IDGenerator(new Lifeline, 'control_no', 4, $year); /** Generate control no */
@@ -294,9 +310,35 @@ class LifelineController extends Controller
     public function destroy(string $id)
     {
         $application = Lifeline::find($id);
-        $application->delete();
+        
 
-        return redirect(route('lifeline.index'))->withSuccess('Record Successfully deleted!');
+        try {
+                $update_account = DB::connection('sqlSrvBilling')
+                    ->table('Consumers Table')
+                    ->where('Accnt No', $application->account_no)
+                    ->update([
+                        'LFflag' => "NO",
+                        'LFdate' => null,
+                        'LFTag' => null,
+                    ]);
+
+                if($update_account == 0){
+                    return redirect()->back()->withError("Invalid Account No.");
+                }
+                else{
+                    $application->delete();
+                    DB::commit();
+                    return redirect(route('lifeline.index'))->withSuccess('Record Successfully deleted!');
+                    // all good
+                }
+                
+            } catch (\Exception $e) {
+                DB::rollback();
+                // something went wrong
+                return redirect()->back()->withError($e->getMessage());
+            }
+
+        //return redirect(route('lifeline.index'))->withSuccess('Record Successfully deleted!');
     }
 
     public function approveLifelineIndex()
@@ -319,6 +361,7 @@ class LifelineController extends Controller
                     ->update([
                         'LFflag' => "NO",
                         'LFdate' => null,
+                        'LFTag' => null,
                     ]);
 
                 if($update_account == 0){
@@ -401,6 +444,7 @@ class LifelineController extends Controller
                     ->update([
                         'LFflag' => "Yes",
                         'LFdate' => $lifeline->date_of_application,
+                        'LFTag' => $lifelines->pppp_id ? "4Ps" : "Non-4Ps",
                     ]);
 
             }
